created xmathutil.hpp
